requirements.txt
apscheduler==3.3.1
boto3==1.4.4
brotlipy==0.7.0
celery==4.0.2
cssselect==1.0.3
dj-database-url==0.4.2
dj-redis-url==0.1.4
django-admin-honeypot==1.0.0
django-cryptography==0.2
django-recaptcha==1.3.1
django-redis==4.8.0
django-storages==1.6.3
django-two-factor-auth==1.6.2
django-waffle==0.12.0
django-widget-tweaks==1.4.1
django==1.11.3
django_otp==0.4.0.1
djangorestframework==3.6.3
furl==1.0.0
geoip2==2.6.0
gunicorn==19.7.1
lxml==4.1.1
pillow==4.2.1
psycopg2==2.7.2
pycountry==17.5.14
pyquery==1.4.0
python-intercom==3.1.0
python-redis-lock==3.2.0
raven==6.1.0
requests==2.18.4
retrying==1.3.3
tqdm==4.19.9
